Move policy specific utilities into a dedicated module.
While iam.py might have been a better home, ec2.py and iam.py are under different licenses
